# Optimize QueryEntry and unify CacheEntry architecture