--- a/lib/log_bench/log/query_entry.rb
+++ b/lib/log_bench/log/query_entry.rb
@@ -14,24 +14,25 @@ module LogBench
       SAVEPOINT = "SAVEPOINT"
       SQL_OPERATIONS = [SELECT, INSERT, UPDATE, DELETE, TRANSACTION, BEGIN_TRANSACTION, COMMIT, ROLLBACK, SAVEPOINT].freeze
 
-      def initialize(raw_line)
-        super
-        self.type = :sql
+      def initialize(raw_line, cached: false)
+        super(raw_line)
+        self.type = cached ? :cache : :sql
+        @cached = cached
       end
 
       def self.build(raw_line)
         return unless parseable?(raw_line)
 
         entry = Entry.new(raw_line)
-        return unless entry.type == :sql
+        return unless [:sql, :cache].include?(entry.type)
 
-        new(raw_line)
+        # Create QueryEntry for both SQL and CACHE entries
+        cached = entry.type == :cache
+        new(raw_line, cached: cached)
       end
 
       def duration_ms
-        return 0.0 unless timing
-
-        timing.gsub(/[()ms]/, "").to_f
+        @duration_ms ||= calculate_duration_ms
       end
 
       def select?
@@ -70,28 +71,24 @@ module LogBench
         operation == SAVEPOINT
       end
 
-      def slow?(threshold_ms = 100)
-        duration_ms > threshold_ms
+      def cached?
+        @cached
       end
 
-      def to_h
-        super.merge(
-          content: content,
-          timing: timing,
-          operation: operation,
-          duration_ms: duration_ms,
-          has_ansi: has_ansi_codes?(content)
-        )
+      def hit?
+        cached? && content.include?("CACHE")
       end
 
       private
+
+      attr_accessor :operation
 
       def extract_from_json(data)
         # Call parent method which checks for request_id
         return false unless super
 
         message = data["message"] || ""
-        return false unless sql_message?(data)
+        return false unless sql_message?(data) || cache_message?(data)
 
         self.content = message.strip
         extract_timing_and_operation
@@ -99,31 +96,48 @@ module LogBench
       end
 
       def extract_timing_and_operation
-        clean_content = remove_ansi_codes(content)
-        self.timing = extract_timing(clean_content)
-        self.operation = extract_operation(clean_content)
+        self.timing = extract_timing
+        self.operation = extract_operation
       end
 
-      def extract_timing(text)
-        match = text.match(/\(([0-9.]+ms)\)/)
+      def extract_timing
+        match = clean_content.match(/\(([0-9.]+ms)\)/)
         match ? match[1] : nil
       end
 
-      def extract_operation(text)
-        SQL_OPERATIONS.find { |op| text.include?(op) }
+      def extract_operation
+        SQL_OPERATIONS.find { |op| clean_content.include?(op) }
       end
 
-      def remove_ansi_codes(text)
-        text.gsub(/\e\[[0-9;]*m/, "")
+      def clean_content
+        @clean_content ||= content&.gsub(/\e\[[0-9;]*m/, "") || ""
       end
 
-      def has_ansi_codes?(text)
-        text.match?(/\e\[[0-9;]*m/)
+      def has_ansi_codes?
+        @has_ansi_codes ||= content&.match?(/\e\[[0-9;]*m/) || false
       end
 
-      private
+      def calculate_duration_ms
+        return 0.0 unless timing
 
-      attr_accessor :operation
+        timing.gsub(/[()ms]/, "").to_f
+      end
+
+      def clear_memoized_values
+        @duration_ms = nil
+        @clean_content = nil
+        @has_ansi_codes = nil
+      end
+
+      def content=(value)
+        super
+        clear_memoized_values
+      end
+
+      def timing=(value)
+        super
+        clear_memoized_values
+      end
     end
   end
 end