… lib/log_bench/log/cache_entry.rb
… lib/log_bench/log/request.rb